Refactor of Network Component
