+++ .env.production
@@ -1,5 +1,5 @@
 # URL to the sleeper backend
-VUE_APP_API_URL=""
+VUE_APP_API_URL="https://sleeper.cozy.fleexo.me"
 
 # JWT Token used for authenticating with the sleeper backend
 VUE_APP_API_JWT_TOKEN=""
